backend/requirements.txt
fastapi
uvicorn
transformers
torch
python-multipart
easyocr
pytesseract
groq
requests
python-dotenv
pytesseract 
pillow